requirements.txt
asgiref==3.2.3
Django>=3.0.3
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.11.0
django-rest-swagger==2.2.0
logzero==1.5.0
